bpsk31: Reduce the width of the symbol filter
A second filter stage has been added, so the symbol filtering
is done at 500Hz sample rate before being resampled to 8KHz.

This dramatically improves performance, and allows this to run in
debug mode without overruns.

The resampling filter is a simple first-order interpolator.
I haven't investigated its effect on the bandwidth/ISI;
I'm sure it could be improved in the future.